+++ core-ng/src/main/java/core/framework/internal/log/InternalActionContext.java
@@ -1,9 +1,0 @@
-package core.framework.internal.log;
-
-/**
- * @author neo
- */
-public final class InternalActionContext {    // use static typing for performance
-    public boolean suppressSlowSQLWarning;
-    public int maxDBOperations;
-}